:recycle: Refactor the user count return value type

+++ src/include/database/dynamo.h
@@ -36,5 +36,5 @@ public:
   int initialize() override;
   int create_user(user) override;
   std::vector<std::string> list_users() override;
-  int count_user() override;
+  int64_t count_user() override;
 };
